fix page fault in rust helper calls
In our current code the rust wrappers for BPF helpers are put at a
lower address than _start in the text section by the compiler
since the compiler followers the source code order. Because we put
the entry point (_start function) at the beginning of the
executable mapping in the kernel, the rust wrappers will be
outside the mapping and we are seeing the page faults upon wrapper
invocation. Add the #[inline(always)] attribute to the wrappers to
ask the compiler to always inline them so that we can avoid this
issue.