inner_unikernel_rt: Implement the new design for tracepoint progs
The new design implements a new prog_type trait with a method that
invokes the program:

pub(crate) trait prog_type {
    fn prog_run(&self, ctx: *const ()) -> u32;
}

The programs are now representated by Rust objects of specific program
types which means the actual code is stored within the object. Each
program types should implement the prog_type trait and the prog_run
method to perform context conversion (if needed) and call the actual
inner-unikernel program.

During CodeGen, a specific LLVM pass will generate the **real**
entrypoint (i.e. the one interfacing with the kernel) that calls the
entry function in this crate -- which in turns calls the prog_run on a
specific program object.

Helper functions would be implemented as methods on these program types
for access control. The support will be added to later commits.

Signed-off-by: Jinghao Jia <jinghao7@illinois.edu>