inner_unikernel_rt: Refactor the helper interface to turn them into methods
Continuing the last commit, turning the helpers into methods in the
implementation of each program type so that access control to helpers
can be implemented.

A possible optimization: since the helpers now have yet another wrapper
call, maybe we can inline the actual helper code?

Also note that bpf_trace_printk cannot be made a method at this time (it
is currently a macro), this is because:
- it uses variadic arguments
- it needs to copy the string slice into a [u8] and append a null
  terminator, which requires a consteval of str.len

Right the inner function bpf_trace_printk_fn is mark as unsafe so that
users cannot arbitrarily call it (assuming we have the forbid(unsafe)
machanism in place.)

We could solve it by adding another C layer between the Rust code and
the kernel code. Since when Rust send &str through the FFI interface it
also sends the length. The intermediate layer could just pad the '\0'
character and call the kernel helper.

Signed-off-by: Jinghao Jia <jinghao7@illinois.edu>