inner_unikernel_rt: Use new bpf_trace_printk_iu implementation
Although there is an upper limit on the size of the format string, it is
not optimal to define a [u8] on the stack with that size (1K). This is
because the kernel stack is only 4K. In the kernel this is defined as a
static variable so that it resides in the data section.

The solution now is to implement a specialized bpf_trace_printk_iu
function in the kernel that defines yet another static 1K character
array. This array will be used for null-terminating the Rust &str. The
array is not dynamically allocated because the eBPF context (usually)
cannot sleep and the GFP_ATOMIC flag has limitations under certain
contexts.

Signed-off-by: Jinghao Jia <jinghao7@illinois.edu>