inner_unikernel_rt: Add RTTI field to prog_type structs
The Rust compiler strips the type information in the generated LLVM IR.
However, the LLVM entry-generating pass needs to know about the type of
the object so that it knows which prog_run method to invoke. Therefore
we add a new RTTI field at the beginning of the struct as an u32 that
stores the BPF_PROG_TYPE_XXX. Since the initialization of the static
program objects are always constexpr, we would be able to read the RTTI
value from the static struct in LLVM. Also, we force the C layout of
structs so that the RTTI is always the first field.

Signed-off-by: Jinghao Jia <jinghao7@illinois.edu>