libiu: skip map entries for relocation
The map pointers are also in the .rela.dyn section. However, these are
handled by the kernel differently because the values will eventually be
the address of the map object in the kernel, rather than *ABS*+value.

Skip the map entries during relocation parsing. This will ensure that
BPF_PROG_LOAD_IU_BASE only sends relocation that needs to be fixed.

Signed-off-by: Jinghao Jia <jinghao7@illinois.edu>