inner_unikernel_rt: Support mut ctx and override_return for kprobe
Supporting bpf_override_return() requires a mutable pt_regs reference.
While we can unsafely convert the immutable ref to a mutable ref, it is
better to this entirely in the safe world. This means the program would
get a &mut pt_regs as the context so it is important to make sure that
user progs cannot change it arbitrarily.

Create a new binding of pt_regs with all fields marked as pub(super),
i.e. only accessible within inner_unikernel_rt::kprobe. On the user
side, programs can use the provided accessors to read register values,
but cannot modify them since no mutators are implemented. The user
should not be able to directly assign this reference a new value either,
given that they will not able to create another instance of pt_regs
(private fields, no pub ctor)

Signed-off-by: Jinghao Jia <jinghao7@illinois.edu>