inner_unikernel_rt: make per-cpu APIs unsafe
The per-cpu APIs really depends on the validity of the specified address
of the per-cpu variable and the specified type to avoid undefined
behavior.

For example:

pub(crate) unsafe fn this_cpu_read<T: PerCPURead<T>>(pcp_addr: u64) -> T {
    <T as PerCPURead<T>>::this_cpu_read(pcp_addr)
}

For this function the compiler will make sure that the type T has an
implementation of PerCPURead<T>, i.e. reading value of T from a per-cpu
variable is supported in general (as an example of T that is not
supported, considering some T where core::mem::size_of::<T>() > 8).

However, the caller has the responsibility to make sure the following
holds:
- pcp_addr must be a static address (as in System.map) of a valid
  per-cpu variable.
- The per-cpu variable with pcp_addr must be a type T

Violating either would result in a UB.

Mark all per-cpu APIs as unsafe.

Signed-off-by: Jinghao Jia <jinghao7@illinois.edu>