inner_unikernel_rt: move to a new map implementation
Previous implementation uses two static variables per map with one
pointing to the other:

  static __MAP1: IUMap<...> = IUMap::new(...);
  static MAP1: &IUMap<...> = &__MAP1;

In the above example the `__MAP1` is the variable that stores all the
map metadata and is parsed by libiu to for map creation. With the new
map created, the value of `MAP1` will be overwritten by libiu to the map
fd (and later to the kernel map pointer by the kernel), i.e. this
point-to relationship no longer holds after this point. However, since
these all happens after compilation, the compiler always assume such
a point-to relationship and it sometimes leads to incorrect optimization
/ code generation (and breaking this assumption is probably by itself an
UB).

Therefore, we are now switching to a new implementation of the maps.
Here we store the kernel side pointer into the `IUMap` struct itself so
that we only need one static variable per map:

  static MAP1: IUMap<...> = IUMap::new(...);

The pointer will be set to null at compile time via the
constantly-evaluated `IUMap::new`. It will be overwritten first by libiu
to the map fd and then by the kernel to the actual kernel map address.

Some other notes:
- The pointer field is set to be crate-private, this prevents the user
  from crating arbitrary pointer values. The pointer can only be `NULL`
  or a valid kernel pointer.
- All the map helpers now requires the `IUMap` struct to have static
  lifetime, this means user cannot use arbitrarily created maps in these
  helpers (e.g. local variables). However, users can still use a static
  `IUMap` not instrumented by libiu and kernel (and therefore has
  a `NULL` kernel map pointer). The helper function will also check the
  validity of the map pointers. Since as mentioned above, the user
  cannot tempt with the pointer value, the `NULL` check is sufficient
  for this.

Signed-off-by: Jinghao Jia <jinghao7@illinois.edu>