inner_unikernel_rt: make access to kernel map ptr volatile
Without the volatile load the compiler is smart enough to use
optimizations (probably SROA and SCCP) to optimize away the entire map
access. This is because from the compiler's perspective the kernel map
ptr is always null and is stored in an immutable static variable.

Use core::ptr::read_volatile to infor the compiler that this is indeed a
volatile variable.

Fixes: e9713e0b10e4 ("inner_unikernel_rt: move to a new map implementation")
Signed-off-by: Jinghao Jia <jinghao7@illinois.edu>