samples/*: compile "core" itself as part of crate compilation
The "core" library built during bootstrap process does not use the
kernel code model and contains instructions that are not supposed to be
used in the kernel (e.g. SIMD instructions).

Use the build-std feature of cargo to compile "core" from source code
using the same profile as our extension program.

Link: https://doc.rust-lang.org/nightly/cargo/reference/unstable.html#build-std
Signed-off-by: Jinghao Jia <jinghao7@illinois.edu>