inner_unikernel_rt: fix incorrect access to __per_cpu_offset
The previous way of accessing the __per_cpu_offset array is incorrectly
treating __per_cpu_offset[0] as the start address of the array (which
should really be &__per_cpu_offset[0] or just &__per_cpu_offset). This
leads to a page fault that cannot be handled when a per-cpu variable is
accessed using the more expensive this_cpu_ptr() / this_cpu_ptr_mut().

Declare the symbol as an extern array in the RT crate so that we can
directly index into it without manually playing with the address.

Fixes: 1d2e0b3d7a8c ("inner_unikernel_rt: declare all used kernel symbols as extern \"C\"")
Signed-off-by: Jinghao Jia <jinghao7@illinois.edu>