inner_unikernel_rt: mark __iu_entry_* functions always-inline
This allows us to eliminate one level of function call in the presence
of LTO on the entry path:

Using samples/hello as example
Before:
  iu_prog1 -> __iu_entry_tracepoint -> hello::hello
After:
  iu_prog1 -> hello::hello

The piece of code that might be duplicated is the __iu_entry_* function,
which should be an acceptable trade off.

Elination of the other level will be come in the next patch.

Signed-off-by: Jinghao Jia <jinghao7@illinois.edu>