samples/*: mark program function as always-inline
Following the previous patch, this allows us to eliminate the other
level of function call on the entry path:

Using samples/hello as example
Before:
  iu_prog1 ->  hello::hello
After:
  iu_prog1

The users can choose to inline the function or not given this is in the
extension program code. It is recommended to enable inlining if the
program function is not used by more than one extension programs.

Signed-off-by: Jinghao Jia <jinghao7@illinois.edu>