librex: use modern attribute syntax
For C++ code (e.g., librex itself), this is the standard way of
specifying attributes. For C code (e.g., the loader programs using the
librex header), although this is supposed to be a C23 (which is not
finalized at this point) feature, both gcc and clang supports it out of
the box even without the '-std=c23' flag. So, let's just use it as it is
more concise.

Signed-off-by: Jinghao Jia <jinghao7@illinois.edu>