rex: make pt_regs a program-type-independent API
The reason for our own API of pt_regs instead of directly exposing the
generated binding is that we would like to prevent unwanted writes to
the registers from user-implemented programs but still be able to modify
the registers in the rex crate (e.g., to support override_return in
kprobe). This means the pt_regs struct cannot be marked as immutable, as
that will prevent modification in rex crate. Therefore, the idea is to
wrap around the kernel pt_regs and implement only accessor methods.

We previously have duplicated APIs in both kprobe and perf event where
the implementation of these APIs is closely coupled with the program
types. It now seems to be a better idea to split pt_regs into its own,
which removes duplicate code and make the API cleaner.

旅に終わりなんてない
夜明けがこない夜もない

Signed-off-by: Jinghao Jia <jinghao7@illinois.edu>