rex/map: fix broken RexRingBuf ABI
Rex maps always need to have the following layout for librex to
recognize them:

  #[repr(C)]
  pub struct RexMapHandle {
      map_type: u32,
      key_size: u32,
      val_size: u32,
      max_size: u32,
      map_flag: u32,
      pub(crate) kptr: *mut (),
      // Some other zero-sized markers
  }

RexRingBuf is previously missing the key_size and val_size fields.
Adding them to be compliant to the ABI.

Side note: RingBufs are still experimental and we currently do not even
have a user in the tree. It is likely the code is still broken somewhere
else.

Fixes: 6b1f7224619e ("rex: update ringbuf API + add test (does not work yet)")
Signed-off-by: Jinghao Jia <jinghao7@illinois.edu>