librex: do not skip .maps relocations
Map fixups are previously handled as special relocations (due to the
need to rewrite to fd / kernel address) and are therefore skipped when
handling other regular relocations.  Since commit a757ddb85ee3 ("libiu:
Support new map data layout"), fixups of kernel map addresses are no
longer handled as relocations, instead, we patch the new kptr member of
the map_def struct during loading. However, the commit left the skipping
of map relocations as it was. This causes problems when the linker
(notably mold) decides to route accesses to the map symbol (i.e., the
map_def struct and the kptr member) through the GOT, as the GOT entry of
the map symbol is left unpatched.

Remove the unused code and always patch .maps relocations.

Fixes: a757ddb85ee3 ("libiu: Support new map data layout")
Signed-off-by: Jinghao Jia <jinghao7@illinois.edu>