rex: impl Drop for AlignedMut to ensure proper writeback
Implement `Drop` for `AlignedMut` to perform automatic writeback to the
original packet (`&mut [c_uchar]`). The `drop` implementation will only
write back to the slice if a value has been copied to avoid unaligned
accesses.

This is the last piece to allow user code to treat `AlignedMut<'_, T>`
as `&mut T`, all the house keeping on the unaligned case (i.e., copy
and writeback) are transparent to user code.

Signed-off-by: Jinghao Jia <jinghao7@illinois.edu>